chore: move customized property to misc
